Accept dictionary as a single argument in `__init__`.
It could happen, that field names would clash with existing
argument (e.g. `self`). In order to not depend on this, this change
is being introduced. Now, to create a Mapper, you have to pass
dictionary of parameters. If you would like to use keyword arguments,
you can use `from_kwargs` factory method.